done : upload image to server & give imagePath to product/category fr…

+++ server/controller/product.controller.js
@@ -1,10 +1,11 @@
-const { Product } = require("../model/product.model");
+const { Product } = require('../model/product.model');
+const fs = require('fs').promises;
 
 async function getAll(req, res) {
   try {
     const products = await Product.find();
 
-    console.log("Found:", products);
+    console.log('Found:', products);
 
     return res.send(products);
   } catch (error) {
@@ -20,25 +21,9 @@ async function getById(req, res) {
   try {
     const product = await Product.findOne({ _id: req.body.id });
 
-    console.log("Found:", product);
+    console.log('Found:', product);
 
     return res.send(product);
-  } catch (error) {
-    console.error(`Error: ${error}`);
-
-    return res
-      .status(500)
-      .send({ error: `Failed to complete the request! Error: ${error}` });
-  }
-}
-
-async function getByCategory(req, res) {
-  try {
-    const products = await Product.find({ category: req.body.category });
-
-    console.log("Found:", products);
-
-    return res.send(products);
   } catch (error) {
     console.error(`Error: ${error}`);
 
@@ -64,7 +49,7 @@ async function add(req, res) {
 
     await product.save();
 
-    return res.send({ ok: "ok" });
+    return res.send({ ok: 'ok' });
   } catch (error) {
     console.error(`Error: ${error}`);
 
@@ -73,15 +58,31 @@ async function add(req, res) {
       .send({ error: `Failed to complete the request! Error: ${error}` });
   }
 }
-
 async function deleteById(req, res) {
   try {
-    await Product.deleteOne({ _id: req.body.id });
+    const product = await Product.findOne({ _id: req.body.id });
+    if (!product) {
+      return res.status(404).send({ error: 'Product not found' });
+    }
 
-    return res.send({ ok: "ok" });
+    const imagePath = product.imgSrc;
+
+    await Product.deleteOne({ _id: req.body.id });
+    if (imagePath) {
+      try {
+        await fs.unlink(imagePath.replace('http://localhost:3000/', ''));
+      } catch (error) {
+        if (error.code === 'ENOENT') {
+          console.log(`File ${imagePath} not found. Continuing...`);
+        } else {
+          throw error;
+        }
+      }
+    }
+
+    return res.send({ ok: 'ok' });
   } catch (error) {
     console.error(`Error: ${error}`);
-
     return res
       .status(500)
       .send({ error: `Failed to complete the request! Error: ${error}` });
@@ -90,6 +91,8 @@ async function deleteById(req, res) {
 
 async function updateById(req, res) {
   try {
+    const product = await Product.findOne({ _id: req.body._id });
+    const oldImagePath = product.imgSrc;
     await Product.updateOne(
       { _id: req.body._id },
       {
@@ -105,7 +108,20 @@ async function updateById(req, res) {
       }
     );
 
-    return res.send({ ok: "ok" });
+    if (oldImagePath) {
+      const imageURL = oldImagePath.replace('http://localhost:3000/', '');
+      try {
+        await fs.unlink(imageURL);
+      } catch (error) {
+        if (error.code === 'ENOENT') {
+          console.log(`File ${imageURL} not found. Continuing...`);
+        } else {
+          throw error;
+        }
+      }
+    }
+
+    return res.send({ ok: 'ok' });
   } catch (error) {
     console.error(`Error: ${error}`);
 
@@ -119,7 +135,7 @@ async function getByCategory(req, res) {
   try {
     const products = await Product.find({ category: req.body.category });
 
-    console.log("Found:", products);
+    console.log('Found:', products);
 
     return res.send(products);
   } catch (error) {
@@ -138,5 +154,4 @@ module.exports = {
   add,
   deleteById,
   updateById,
-  getByCategory,
 };
